+++ tests/novax/api/test_nats.py
@@ -2,10 +2,11 @@ import asyncio
 
 import pytest
 
+from nova.core.nats import Message
 from nova.core.nova import Nova
-from nova.events.nats import Message
 
 
+@pytest.mark.xdist_group("program-runs")
 @pytest.mark.asyncio
 async def test_nats_pub_sub():
     nova = Nova()
@@ -18,11 +19,9 @@ async def test_nats_pub_sub():
         nonlocal collected_message
         collected_message = msg
 
-    await nova.api_gateway.subscribe("nova.test.subject", cb=cb)
+    await nova.api_gateway.subscribe("nova.test.subject", on_message=cb)
     nova.api_gateway.publish_message(Message(subject="nova.test.subject", data=b"test message"))
 
-    # todo, wait for message, this makes the test flaky
-    # but since this is integration test, it is okay for now
     await asyncio.sleep(2)
 
     assert collected_message is not None, "No message received"
